+++ fbw-a32nx/src/base/flybywire-aircraft-a320-neo/SimObjects/AirPlanes/FlyByWire_A320_NEO/gsx_handler.py
@@ -1,40 +1,8 @@
 friendlyName = "FlyByWire A32NX Handler"
+# Ensuring that Chocks and Brakes Tests are not overriden by an User-Profile
 chocksTest = "(L:A32NX_PARK_BRAKE_LEVER_POS,number) (L:A32NX_GND_EQP_IS_VISIBLE,number) ||"
 brakesTest = "(L:A32NX_PARK_BRAKE_LEVER_POS,number)"
-zfw = LVariable("A32NX_AIRFRAME_ZFW")
-
-def onAircraftEngaged(self):
-    pass
-
-def onAircraftDisengaged(self):
-    pass
-
-def onBeforeVehicleSelect(self):
-    pass
-
-def onBoardingRequested(self):
-    pass
-
-def onDeboardingRequested(self):
-    pass
-
-def onDepartureRequested(self):
-    pass
-
-def onCateringRequested(self):
-    pass
-
-def onWaterServiceRequested(self):
-    pass
-
-def onWaterServiceCompleted(self):
-    pass
-
-def onLavatoryServiceRequested(self):
-    pass
-
-def onDeicingAction(self):
-    pass
+fuelSync = LVariable("A32NX_GSX_FUEL_SYNC_ENABLED")
 
 def onJetwayConnected(self):
     pass
@@ -42,47 +10,23 @@ def onJetwayConnected(self):
 def onJetwayDisconnected(self):
     pass
 
-def onBypassPinConnected(self):
-    pass
-
-def onBypassPinDisconnected(self):
-    pass
-
-def onWaterOperateDoor(self, value=True):
-    pass
-
-def onLavatoryOperateDoor(self, value=True):
-    pass
-
-def operateChocks(self):
-    pass
-
-def operateGpu(self):
-    pass
-
 def setExtPowerAvail(self):
     pass
 
-def setFuelLevel(self, level):
-    pass
-
-def setPlannedFuel(self):
-    pass
-
-def getCurrentZFW(self):
-    return zfw.getValue()
-
-def fuelPumpInsist(self):
-    return 1
-
-def setPayload(self):
-    pass
-
-def removePayload(self):
-    pass
-
 def customTruckRequestedMessage(self):
-    pass
+    if fuelSync.getValue() != 0:
+        return "Fuel Truck requested - load OFP/set planned Fuel in FlyPad"
+    else:
+        return self._super_customTruckRequestedMessage()
+
+def customRefuelMessage(self):
+    if fuelSync.getValue() != 0:
+        return "GSX Refuel in Progress"
+    else:
+        return self._super_customRefuelMessage()
 
 def customTruckInPositionMessage(self):
-    pass
+    if fuelSync.getValue() != 0:
+        return "Fuel Truck in Position - Refuel starts automatically"
+    else:
+        return self._super_customTruckInPositionMessage()