+++ .github/CHANGELOG.md
@@ -78,7 +78,8 @@
 1. [FMS] Fixed incorrect course for enroute holds and direct tos - @tracernz (Mike)
 1. [FMS] Show true bearings correctly in various FMS pages on the MCDU/MFD - @tracernz (Mike)
 1. [A380X/FWS] Improve various ABN SENSED procedure accuracy - @Jonny23787 (Jonathan)
-1. [GSX] Changes to support new Loader Behavior, Handlers and Service Cancellation for A380X/A32NX - @Fragtality (Fragtality)
+1. [GSX] Changes to support new Features: Loader Behavior, Python Handlers and Service Cancellation - @Fragtality (Fragtality)
+1. [GSX] Reworked Ground Power Synch with Jetway/GSX GPU - @Fragtality (Fragtality)
 
 ## 0.14.0